Add minigallina
This provides an experimental gallina-like language to define constrs
which simplifies writing terms for tests. The language is a subset of
Coq's gallina and considerably more explicit:

- no type inference, everything must be specified explicitly
- no notations
- no multi pattern matches
- global names must be specified by fully qualified path

The API as well as the supported Gallina subset is subject to change,
trying out how much this helps for writing other unit tests.