minigallina: correctly record number of match arguments
The number of match arguments was accidentally always set to zero for
branches of match expression (due to moving the formal argument list before
taking the count). Fix this and adapt the unit test.

The number of arguments is not quite consistent withwhat coq emits: Coq will
omit the parametrisation of the inductive type from match arguments. However
for the moment this is good enough to be accurately parseable.