fix de bruijn indices for fixpoint expressions
The type signature for fixpoint functions in fixpoint expressions is
represented in two ways:
1. as a product describing the signature, resolved _without_ adding
   the function signatures to the local context
2. as lambdas with formal arguments and result expression, resolved _with_
   adding the function signatures to the local context

This resulted in an inconsistent specification of the function return type
(using method 1) with respect to the formal arguments of the function
(using method 2 above). This change restores consistenty by shifting
the return type so it is also represented using method 2.